refactor: 추상클래스로 일부 메서드/변수 이동
