+++ src/main/java/blackjack/domain/human/Human.java
@@ -5,14 +5,19 @@ import blackjack.domain.card.Cards;
 import blackjack.util.Constants;
 
 public abstract class Human {
+    protected final Name name;
+    protected final Cards cards;
 
-    public abstract String getName();
-
-    public abstract void addCard(final Card card);
-
-    public abstract Cards getCards();
+    protected Human(Cards cards, String name) {
+        this.name = Name.of(name);
+        this.cards = cards;
+    }
 
     public abstract boolean isAbleToHit();
+
+    public void addCard(final Card card) {
+        cards.add(card);
+    }
 
     public int getPoint() {
         return getCards().getPoint();
@@ -20,5 +25,13 @@ public abstract class Human {
 
     public boolean isTwoCard() {
         return getCards().size() == Constants.INIT_CARD_NUMBER;
+    }
+
+    public String getName() {
+        return name.get();
+    }
+
+    public Cards getCards() {
+        return cards;
     }
 }
