+++ src/main/java/blackjack/domain/human/Name.java
@@ -26,11 +26,4 @@ public class Name {
     public String get() {
         return name;
     }
-
-    @Override
-    public String toString() {
-        return "Name{" +
-                "name='" + name + '\'' +
-                '}';
-    }
 }
